add pics for ae60d3, fix 4b9030 in plane_images.csv (#399)
add pics for ae60d3.

Also realized that fixing the capitalization of 4B9030 in plane-alert-db.csv did not propagate to plane_images.csv, but rather, it created a new entry for it while preserving the 4b9030 entry. So I just deleted the empty one and fixed the capitalization on the extant one with pics.